Add files via upload
A dynamic and interactive calculator built using HTML, CSS, and JavaScript, combining the power of these web technologies to create a seamless and user-friendly mathematical tool. The HTML structure defines the calculator's layout, with buttons and display areas, while CSS styles enhance the visual appeal, providing a clean and modern interface. JavaScript handles the logic behind the calculations, enabling real-time updates as users input numbers and operations. This calculator not only performs basic arithmetic but also showcases the synergy between HTML for structure CSS for aesthetics, and JavaScript for functionality, delivering a responsive and intuitive calculator experience for users.